refactor: User 엔티티 변경에 따른 User 관련 코드 수정 및 User 서비스 로직 개선
