+++ moamoa/src/main/java/com/likelion/moamoa/domain/user/repository/UserRepository.java
@@ -1,0 +1,15 @@
+package com.likelion.moamoa.domain.user.repository;
+
+import com.likelion.moamoa.domain.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    boolean existsByEmail(String email);
+
+    Optional<User> findByEmail(String email);
+}
